Update Go version in deploy script

+++ k9s/deploy.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-GOVERSION=1.22.2
+GOVERSION="1.22.2"
 
 # Stop execution on any error
 set -e
@@ -9,42 +9,35 @@ set -e
 sudo apt-get update
 sudo apt-get install -y build-essential
 
-# Go to the user's home directory
-cd ~
-
-# Download Go binary
-curl -OL https://go.dev/dl/go${GOVERSION}.linux-amd64.tar.gz
-
-# Remove any existing Go installation and extract the new version
-sudo rm -rf /usr/local/go
-sudo tar -C /usr/local -xzf go${GOVERSION}.linux-amd64.tar.gz
-
-# Add Go to PATH for all future terminal sessions
-export PATH=$PATH:/usr/local/go/bin
-
-# Check if Go PATH is already in .bashrc and add it if not
-if ! grep -q '/usr/local/go/bin' ~/.bashrc; then
-    echo "export PATH=\$PATH:/usr/local/go/bin" >> ~/.bashrc
+if ! command -v go version &> /dev/null; then
+  cd ~
+  curl -OL https://go.dev/dl/go${GOVERSION}.linux-amd64.tar.gz
+  sudo rm -rf /usr/local/go
+  sudo tar -C /usr/local -xzf go${GOVERSION}.linux-amd64.tar.gz
+  export PATH=$PATH:/usr/local/go/bin
+  if ! grep -q '/usr/local/go/bin' ~/.bashrc; then
+      echo "export PATH=\$PATH:/usr/local/go/bin" >> ~/.bashrc
+  fi
+  go version
+  source ~/.bashrc
+  echo -e " \033[32;5mGo installation complete\033[0m"
+else
+  go version
+  echo -e " \033[32;5mGo already installed\033[0m"
 fi
-
-# Load the new PATH into the current shell session
-source ~/.bashrc
-
-# Verify Go installation
-go version
 
 # Check if k9s repo is already cloned
 if [ -d "~/k9s" ]; then
     # Repo exists, so pull any new changes
-    cd k9s
+    cd ~/k9s
     git pull
 else
     # Repo does not exist, clone it
     git clone https://github.com/derailed/k9s.git
-    cd k9s
+    cd ~/k9s
 fi
 
-
 # Build k9s and run it
+cd ~/k9s
 make build
 sudo cp ./execs/k9s /usr/local/bin
